PC-1228: split the address manual pages into three
this means that there can be distinct logic per address manual page

in this case, on entering an address the fact the user did this is saved to the session. routing functions can then know at which point the user decided to enter an address

while on the address pages the route is temporarily invalid, so we have to generate back url manually. this is not great, but it resolves itself after inputting a manual address